Support for double precision on CPU and GPU

+++ crow/crow/transfer/cputransfer.py
@@ -5,18 +5,19 @@ from mpi4py import MPI
 EPSILON = 10**(-9)
 
 comm = MPI.COMM_WORLD
+#dtype = np.float64
 
-def npzeros(n, m):
-    return np.array(np.zeros((n, m)), dtype=np.float32, order='C')
+def npzeros(n, m, dtype=np.float32):
+    return np.array(np.zeros((n, m)), dtype=dtype, order='C')
 
-def npzero_slices(slices, k):
-    return [npzeros(b-a,k) for a,b in slices]
+def npzero_slices(slices, k, dtype=np.float32):
+    return [npzeros(b-a,k, dtype=dtype) for a,b in slices]
 
-def npones(n, m):
-    return np.array(np.ones((n, m)), dtype=np.float32, order='C')
+def npones(n, m, dtype=np.float32):
+    return np.array(np.ones((n, m)), dtype=dtype, order='C')
 
-def npnumber(x):
-    return np.array(np.zeros((1,1))+x, dtype=np.float32, order='C')
+def npnumber(x, dtype=np.float32):
+    return np.array(np.zeros((1,1))+x, dtype=dtype, order='C')
 
 ### Reducer and synchronizations ###
 
@@ -27,7 +28,7 @@ def cpu_reduce(rank, device_list, output, source=0):
     if rank == source:
         kmp = None
         if len(device_list) > 0:
-            kmp = npzeros(*output.shape)
+            kmp = npzeros(*output.shape, dtype=output.dtype)
         for b in device_list:
             cpu_recv(kmp, b)
             output += kmp
